+++ app/src/main/java/com/mezon/mobile/network/MezonApi.kt
@@ -1,7 +1,5 @@
 package com.mezon.mobile.network
 
-import com.mezon.mobile.util.Base58
-
 import com.mezon.mobile.BuildConfig
 import android.util.Base64
 import com.mezon.mezon.api.Account
@@ -60,22 +58,6 @@ import kotlinx.serialization.Serializable
 import javax.inject.Inject
 import javax.inject.Singleton
 class UnauthorizedException(message: String) : RuntimeException(message)
-
-@Serializable
-data class MmnGetAccountResponse(
-    val address: String = "",
-    val balance: String = "0",
-    val nonce: Int = 0,
-    val decimals: Int = 6
-)
-
-@Serializable
-data class MmnJsonRpcResponse(
-    val jsonrpc: String = "2.0",
-    val result: MmnGetAccountResponse? = null,
-    val error: kotlinx.serialization.json.JsonElement? = null,
-    val id: Long = 1
-)
 
 @Serializable
 data class AuthEmailBody(
@@ -663,31 +645,6 @@ class MezonApi @Inject constructor(
             throw RuntimeException("File upload failed (${response.status.value})")
         }
     }
-
-    suspend fun getWalletBalance(userId: String): MmnGetAccountResponse? {
-        try {
-            val address = calculateMmnAddress(userId)
-            val requestBody = "{\"jsonrpc\":\"2.0\",\"method\":\"account.getaccount\",\"params\":{\"address\":\"$address\"},\"id\":1}"
-            val mmnUrl = BuildConfig.MEZON_MMN_API_URL
-            val response = httpClient.post(mmnUrl) {
-                contentType(ContentType.Application.Json)
-                setBody(requestBody)
-            }
-            if (!response.status.isSuccess()) return null
-            val responseBody = response.bodyAsText()
-            val rpcResult = kotlinx.serialization.json.Json { ignoreUnknownKeys = true }.decodeFromString<MmnJsonRpcResponse>(responseBody)
-            return rpcResult.result
-        } catch (e: Exception) {
-            return null
-        }
-    }
-
-    private fun calculateMmnAddress(userId: String): String {
-        val md = java.security.MessageDigest.getInstance("SHA-256")
-        val hash = md.digest(userId.toByteArray(Charsets.UTF_8))
-        return Base58.encode(hash)
-    }
-
 
 }
